Src: Reorganize structure of the project
