+++ src/dialog/OperatorEditorDialog.cpp
