+++ src/dialog/OperatorEditorDialog.hpp
