+++ src/model/CountriesOperatorsModel.cpp
