+++ src/model/TreeIconTextDelegate.hpp
