+++ src/view/OperatorEditor.hpp
