+++ src/view/buttons/AddOperatorButton.hpp
